extract argv shit b/c it's just a lot to slam into a Config value
